pyproject.toml
[project]
name = "encoded-cot"
version = "0.1.0"
description = "Experiments on encoded reasoning with ciphers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0",
    "datasets>=2.14.0",  # For loading MATH 500 dataset
    "openai>=1.0",
    "anthropic>=0.18",
    # For plotting:
    # "matplotlib>=3.8",
    # "seaborn>=0.13",
    # "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pyright>=1.1.0",
    "pytest>=7.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/encoded_reasoning"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pyright]
pythonVersion = "3.13"
